# Sequentialize GetAllUsesOfAllSymolsInFile