--- a/src/fsharp/service/FSharpCheckerResults.fsi
+++ b/src/fsharp/service/FSharpCheckerResults.fsi
@@ -210,10 +210,7 @@ type public FSharpCheckFileResults =
     member GetFormatSpecifierLocationsAndArity : unit -> (range*int)[]
 
     /// Get all textual usages of all symbols throughout the file
-    member GetAllUsesOfAllSymbolsInFile : ?cancellationToken: CancellationToken -> FSharpSymbolUse[]
-
-    /// Get all textual usages of all symbols throughout the file that match the given predicate
-    member GetAllUsesOfAllSymbolsInFileByPredicate : predicate: (FSharpSymbolUse -> bool) * ?cancellationToken: CancellationToken -> FSharpSymbolUse[]
+    member GetAllUsesOfAllSymbolsInFile : ?cancellationToken: CancellationToken -> seq<FSharpSymbolUse>
 
     /// Get the textual usages that resolved to the given symbol throughout the file
     member GetUsesOfSymbolInFile : symbol:FSharpSymbol * ?cancellationToken: CancellationToken -> FSharpSymbolUse[]