… tests/FSharp.Compiler.Service.Tests/SurfaceArea.netstandard.fs
… tests/benchmarks/Benchmarks.sln
… tests/benchmarks/CompilerServiceBenchmarks/CompilerServiceBenchmarks.fsproj
… tests/benchmarks/CompilerServiceBenchmarks/Program.fs
… tests/benchmarks/CompilerServiceBenchmarks/decentlySizedStandAloneFile.fsx
… tests/service/Common.fs
… tests/service/EditorTests.fs
… tests/service/ProjectAnalysisTests.fs
… tests/service/Symbols.fs
… vsintegration/src/FSharp.Editor/CodeFix/SimplifyName.fs
… vsintegration/src/FSharp.Editor/Common/Extensions.fs
… vsintegration/src/FSharp.Editor/Diagnostics/SimplifyNameDiagnosticAnalyzer.fs
… vsintegration/src/FSharp.Editor/Diagnostics/UnusedDeclarationsAnalyzer.fs
… vsintegration/src/FSharp.Editor/Diagnostics/UnusedOpensDiagnosticAnalyzer.fs