~~Some cleanup + a slightly different way to get symbols yielded some minor improvements as per the benchmark that comes with these changes.~~

This sequentializes `GetAllUsesOfAllSymbols`. Moving to a sequence instead of a generated array appears to give, overall, some improvements as measured by a benchmark in the three primary routines in Visual Studio that use this API.

CPU time:

|Routine|Current mean time|New mean time|Result|
|-------|------------|--------|--------|
|Simplify Names|29,670.7 us| 30,076.3 us|1.37% slower|
|Unused Opens|1,408.4 us|1,320.1 us|4.52% **faster**|
|Unused Declarations|730.9 us|250.3 us|65.7% **faster**|

Memory Usage:

|Routine|Current memory usage|New memory usage|Result|
|-------|------------|--------|--------|
|Simplify Names|16263.42 KB|16085.83|1.1% **less** memory|
|Unused Opens|903.31 KB|876.64 KB|3% **less** memory|
|Unused Declarations|749.99 KB|450.82 KB|40% **less** memory|

Since Unused Opens and Unused Declarations run by default after every time we check a document (which is all the damn time), this should improve things a bit. Simplify Names is slightly slower, but the memory usage is also slightly better, and it's off by default. Frankly, Simplify Names needs to be re-architected anyways.

One thing that became apparent to me when writing a first pass of this with `GetAllUsesOfAllSymbolsInFileByPredicate` is that these kinds of symbol operations aren't working with the right data. Every time this is called we traverse a big set of `TcSymbolUseData [] []` and construct `FSharpSymbol`s each time. The `FSharpSymbol` construction is expensive and we need to do it for each item until we get to filter it out. The API could change to work off of the `Item` type, but that would be more complicated and only serve to further push Name Resolution data structures out further into the service layers, which I'd personally like to avoid.

I wonder if it would be better to keep around a pool of `FSharpSymbol`s constructed from the `TcSymbolUseData` instead. That way subsequent calls to this method (and others like it) won't unnecessarily re-create `FSharpSymbol`s all the time. I imagine it might be tricky keeping that resident in memory and "smart" so that we're not re-creating everything all the time, though.

``` ini

BenchmarkDotNet=v0.11.3, OS=Windows 10.0.19042
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit LegacyJIT-v4.8.4250.0 DEBUG
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4250.0
```

**Current**

|             Method |        Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------- |------------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|      SimplifyNames | 29,670.7 us | 376.91 us | 334.12 us |   3906.2500 |    125.0000 |           - |         16263.42 KB |
|        UnusedOpens |  1,408.4 us |  17.34 us |  15.37 us |    193.3594 |     56.6406 |           - |           903.31 KB |
| UnusedDeclarations |    730.9 us |  16.90 us |  26.31 us |    169.9219 |     48.8281 |           - |           749.99 KB |

**Modified**

|             Method |        Mean |      Error |     StdDev |      Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------- |------------:|-----------:|-----------:|------------:|------------:|------------:|------------:|--------------------:|
|      SimplifyNames | 30,076.3 us | 561.798 us | 498.019 us | 29,994.2 us |   3906.2500 |           - |           - |         16085.83 KB |
|        UnusedOpens |  1,344.6 us |  25.686 us |  24.027 us |  1,337.6 us |    212.8906 |      1.9531 |           - |           876.64 KB |
| UnusedDeclarations |    250.3 us |   4.997 us |   9.864 us |    245.5 us |    109.8633 |           - |           - |           450.82 KB |