@cartermp 2020-10-31
@dsyme this is ready for review again. Sequentializing `GetAllUsesOfAllSymbols` offers a substantial perf benefit for finding all unused declarations, a marginal benefit for finding all unused opens, and a marginal detriment to name simplification (which frankly needs a complete re-think anyways). Memory usage is also down across the board.