@auduchinok 2020-11-05
@cartermp Thanks for letting me know!

Related thing: I'm going to revisit [this PR](https://github.com/dotnet/fsharp/pull/8828) about scoping captured types, and the ultimate goal for me there is to try to do the same for symbols, so range-based chunks could be processed independently.