@dsyme 2020-11-06
> I wonder if it would be better to keep around a pool of FSharpSymbols constructed from the TcSymbolUseData instead. That way subsequent calls to this method (and others like it) won't unnecessarily re-create FSharpSymbols all the time. I imagine it might be tricky keeping that resident in memory and "smart" so that we're not re-creating everything all the time, though.

The main problem with this is that it breaks the current layering in the compiler - where FSharpSymbol is used for "public" views and Item etc. is used internally

This is a common problem when trying to reduce allocations - the objects for exterior wrappers need to be pushed down a layer.  

An alternative is to simply use the exterior wrappers within more and more of your codebase.  We could gradually move to doing that, pushing "symbols.fs/fsi" down to just after TypedTree.fs and then using them.  They are however a little heavier as they capture a number of things (TcGlobals etc.) that normally get passed around or picked up from some other context