@auduchinok 2020-11-06
> The main problem with this is that it breaks the current layering in the compiler - where FSharpSymbol is used for "public" views and Item etc. is used internally

As someone who uses `FSharpSymbol` all the time (unlike, say, `FSharpExpr` that we don't use at all), I often want to be able to go down the level, since we're not only accessing symbols but also getting lot of info about each such symbol in many places, making even more objects we don't need right afterwards (consider getting curried parameter groups on `mfv` when all we need is groups count in some place). I'm OK with it being more difficult to work with and possible changing often, but we got used to it after working with `SyntaxTree` over the last few years. `FSharpSymbol` sometimes also feels limited compared to what the compiler has internally.

I'd also be happy to use the internal helpers for reasoning about code instead of ones that work with `FSharpSymbol` as well.